Se cambia algunos valores de la clase y la base de datos

+++ src/main/java/com/ProyectoIntegrador/sistematransaccionesbancarias/infraestructure/data/jpaRepositories/cuenta/CuentaJPARepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CuentaJPARepository extends CrudRepository<CuentaJPAEntity,Integer>{
+public interface CuentaJPARepository extends CrudRepository<CuentaJPAEntity,Long>{
 
     // hay que traer la cuenta por id de un usuario
 }
